Fiddling with converter package layout
The interface is a bit overkill. Yet I think it'll be a blessing to think
ahead for once.
